# src/presentation/http/opt_out_reason_handler.rs
//! OptOutReason REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{OptOutReasonService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateOptOutReasonDto, UpdateOptOutReasonDto, PatchOptOutReasonDto, OptOutReasonResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum OptOutReasonError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for OptOutReasonError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for OptOutReasonError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "OPTOUTREASON_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "OPTOUTREASON_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "OPTOUTREASON_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "OPTOUTREASON_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for OptOutReason.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /opt_out_reasons | List with pagination |
/// | POST | /opt_out_reasons | Create new |
/// | GET | /opt_out_reasons/:id | Get by ID |
/// | PUT | /opt_out_reasons/:id | Full update |
/// | PATCH | /opt_out_reasons/:id | Partial update |
/// | DELETE | /opt_out_reasons/:id | Soft delete |
/// | POST | /opt_out_reasons/bulk | Bulk create |
/// | POST | /opt_out_reasons/upsert | Upsert |
/// | GET | /opt_out_reasons/trash | List deleted |
/// | POST | /opt_out_reasons/:id/restore | Restore |
/// | DELETE | /opt_out_reasons/empty | Empty trash |
/// | GET | /opt_out_reasons/:id/deleted | Get deleted by ID |
/// | DELETE | /opt_out_reasons/trash/:id | Permanent delete from trash |
/// | GET | /opt_out_reasons/count | Count active entities |
/// | GET | /opt_out_reasons/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(OptOutReasonService::with_repository(repository));
/// let router = create_opt_out_reason_routes(service);
/// ```
pub fn create_opt_out_reason_routes(service: Arc<OptOutReasonService>) -> Router {
    BackboneCrudHandler::<OptOutReasonService, OptOutReason, CreateOptOutReasonDto, UpdateOptOutReasonDto, OptOutReasonResponseDto>::routes(
        service,
        "/opt_out_reasons",
    )
}

/// Create Axum router with only the read (GET) endpoints for OptOutReason.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_opt_out_reason_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_opt_out_reason_read_routes(service: Arc<OptOutReasonService>) -> Router {
    BackboneCrudHandler::<OptOutReasonService, OptOutReason, CreateOptOutReasonDto, UpdateOptOutReasonDto, OptOutReasonResponseDto>::read_routes(
        service,
        "/opt_out_reasons",
    )
}

/// Create Axum router with only the write (mutation) endpoints for OptOutReason.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_opt_out_reason_write_routes(service: Arc<OptOutReasonService>) -> Router {
    BackboneCrudHandler::<OptOutReasonService, OptOutReason, CreateOptOutReasonDto, UpdateOptOutReasonDto, OptOutReasonResponseDto>::write_routes(
        service,
        "/opt_out_reasons",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_opt_out_reason_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<OptOutReasonService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_opt_out_reason_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}